requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.1.0
Flask-CORS==5.0.1
Flask-Login==0.6.3
PyMySQL==1.1.1
python-dotenv==1.1.0
python-dateutil==2.9.0
requests==2.32.3
Werkzeug==3.1.3
